+++ scripts/public-additional-scripts.js
@@ -12,7 +12,7 @@ import './public/rebate-query-tool';
 
 import './public/cleanbc/accessibility';
 import './public/betterhomes/accessibility';
-// import './public/betterhomes/rebates-archive';
+import './public/betterhomes/rebates-archive';
 import './public/betterhomes/definitions-form-resubmit';
 import './public/betterbuildings/accessibility';
 import './public/goelectric/accessibility';